+++ src/wallet/utils/verify.ts
@@ -43,7 +43,6 @@ export function verifyArbitrary({
   type = "secp256k1",
 }: VerifyArbitraryParams): boolean {
   const params = {
-    wallet,
     pubKey: base64.decode(pubKey),
     bech32Prefix,
     data: utf8.decode(data),
@@ -54,14 +53,10 @@ export function verifyArbitrary({
     switch (wallet) {
       case WalletName.STATION:
         return verifyECDSA(params);
-      case WalletName.COMPASS:
-      case WalletName.COSMOSTATION:
-      case WalletName.KEPLR:
-      case WalletName.LEAP:
-      case WalletName.NINJI:
-        return verifyADR36(params);
       case WalletName.METAMASK_INJECTIVE:
         return verifyEIP191(params);
+      default:
+        return verifyADR36(params);
     }
   } catch (err) {
     return false;