+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## `v0.0.63`
+
+### Fixes
+
+- Ignore all errors in retrieving public keys for Station Extension
+
+## `v0.0.62`
+
+### Improvements
+
+- Txs are now broadcasted directly via the user-specified RPC instead of via the wallet (closes <https://github.com/coinhall/cosmes/issues/17>)
+- Update Station extension to use Keplr's API interface
+
+## `v0.0.60`
+
+### Features
+
+- Parse `EthermintAccount` in `toBaseAccount` util function
+
+## `v0.0.59`
+
+### Fixes
+
+- Fix Station mobile wallet to not throw errors if pub keys do not exist
+
+## `v0.0.58`
+
+### Fixes
+
+- Fix Station mobile wallet connect issue when connecting more than 1 chain
+
+## `v0.0.57`
+
+### Fixes
+
+- Allow Terra testnet to work correctly with Station extension
+
 ## `v0.0.56`
 
 ### Features